Move guest and staff functions from peep.cpp
